requirements.txt
# Core ML and VLM dependencies
torch>=2.0.0
transformers==4.51.3
accelerate>=0.20.0
sentencepiece>=0.1.97
protobuf>=3.20.0

# UI and web interface
gradio>=4.0.0

# Image processing
pillow>=9.0.0

# HuggingFace Spaces support
spaces>=0.19.0

# OCR alternatives and utilities
easyocr>=1.7.0
opencv-python-headless>=4.5.0

# SSL and networking
certifi>=2021.0.0
urllib3>=1.26.0

# Additional utilities
numpy>=1.21.0
requests>=2.25.0
einops>=0.6.0
timm>=0.9.0 